pyproject.toml
[project]
name = "companion"
version = "0.1.0"
description = "Companion Tools - Collection of content gathering and processing utilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.2.1",
    "crewai-tools>=1.2.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    # Workspace packages - install all tools by default
    "gmail-collector",
    "reddit-collector",
    "stackexchange-collector",
    "web-articles-collector",
    "web-direct-collector",
    "transcribe",
    "pdf-to-md",
]

[tool.uv.workspace]
members = [
    "tools/gmail",
    "tools/reddit",
    "tools/stackexchange",
    "tools/web_articles",
    "tools/web_direct",
    "tools/transcribe",
    "tools/pdf-to-md",
]

[tool.uv.sources]
gmail-collector = { workspace = true }
reddit-collector = { workspace = true }
stackexchange-collector = { workspace = true }
web-articles-collector = { workspace = true }
web-direct-collector = { workspace = true }
transcribe = { workspace = true }
pdf-to-md = { workspace = true }

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
